Copied (everything for) 'Elapsed/WPM' from Mldvp
This means we now have various duplicate functions, an 'Mldvp_' version
and a 'DTScan_' version, notable for CurLoc_Save, CurLoc_Update,
CallerFuncName, and various datetime related functions.

Functionality removed (at any rate) from Mldvp, to be maintained
seperately, (meaning Mldvp can die a little sooner?) and it *appears* to
work in its new home. To be tidied/consoldated at some later point.

dtscan *isn't* being used in these functions -> we are searching for
datetimes using regex from vimscript instead of python -> are there
enough datetimes in a 'paragraph' for it to be likely to matter <-
python does this searching a lot faster.